tools/requirements.txt
GitPython ~= 3.1.27
yaclog ~= 1.0.4
jinja2 ~= 3.1.0